backend/requirements.txt
# BATUHAN — Backend Dependencies
# Install with: pip install -r requirements.txt

# -----------------------------------------------------------------------
# Web Framework
# -----------------------------------------------------------------------
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9        # file upload support
jinja2==3.1.4                  # HTML template rendering for internal UI

# -----------------------------------------------------------------------
# AI Provider
# -----------------------------------------------------------------------
anthropic==0.34.2

# -----------------------------------------------------------------------
# Config & Validation
# -----------------------------------------------------------------------
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# -----------------------------------------------------------------------
# Document Parsing
# -----------------------------------------------------------------------
pdfplumber==0.11.4             # PDF text extraction
PyMuPDF==1.24.10               # PDF fallback + image extraction
python-docx==1.1.2             # DOCX parsing and assembly

# -----------------------------------------------------------------------
# OCR
# -----------------------------------------------------------------------
pytesseract==0.3.13            # Tesseract OCR wrapper
Pillow==10.4.0                 # Image handling

# -----------------------------------------------------------------------
# Background Jobs
# -----------------------------------------------------------------------
celery==5.4.0
redis==5.1.0

# -----------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------
sqlalchemy==2.0.35
aiosqlite==0.20.0              # async SQLite support

# -----------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------
httpx==0.27.2                  # async HTTP client
python-jose==3.3.0             # JWT (future auth)
aiofiles==24.1.0               # async file I/O

# -----------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2                  # test client for FastAPI

